user cannot like post more than one time, user cannot manipulate posts that are not his, small fixes

+++ backend/src/modules/posts/services/DeletePostService.js
@@ -5,11 +5,18 @@ class DeletePostService {
     this.postsRepository = postsRepository;
   }
 
-  async execute({ post_id }) {
+  async execute({ post_id, user_id }) {
     const post = await this.postsRepository.findById(post_id);
 
     if (!post) {
       throw new AppError({ status: 404, error: 'Cannot find post' });
+    }
+
+    if (post.user_id !== user_id) {
+      throw new AppError({
+        status: 401,
+        error: 'You do not have permission to edit this post',
+      });
     }
 
     await this.postsRepository.delete(post);
